HudumaMMU/requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2==2.7.5
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
